Deprecate `edit` API.
